[outpolic-KJW] detail:
