+++ outpolic/src/main/java/outpolic/enter/outsourcingRequest/service/OutsourcingRequestService.java
@@ -1,8 +1,7 @@
-/*
- * package outpolic.enter.outsourcingRequest.service;
- * 
- * import outpolic.enter.outsourcingRequest.domain.OutsourcingRequest;
- * 
- * public interface OutsourcingRequestService { void
- * addOutsourcingRequest(OutsourcingRequest outsourcingRequest); }
- */
+package outpolic.enter.outsourcingRequest.service; // ✅ 패키지 이름 수정
+
+import outpolic.enter.outsourcingRequest.domain.OutsourcingRequest;
+
+public interface OutsourcingRequestService {
+    void addOutsourcingRequest(OutsourcingRequest request);
+}
