+++ outpolic/src/main/java/outpolic/enter/outsourcingRequest/service/impl/OutsourcingRequestServiceImpl.java
@@ -6,30 +6,36 @@ import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
 import outpolic.enter.outsourcingRequest.domain.OutsourcingRequest;
 import outpolic.enter.outsourcingRequest.mapper.OutsourcingRequestMapper;
+import outpolic.enter.outsourcingRequest.service.OutsourcingRequestService;
 
 
-/*
+
 @Service
 @RequiredArgsConstructor
 public class OutsourcingRequestServiceImpl  implements OutsourcingRequestService {
+	
 	private final OutsourcingRequestMapper requestMapper;
 	
 	@Override
-	@Transactional
-	public void addOutsourcingRequest(OutsourcingRequest  outsourcingRequest) {
+	@Transactional // 모든 DB 작업이 하나의 묶음으로 처리되도록 보장
+	public void addOutsourcingRequest(OutsourcingRequest request) {
+		// 1. 새로운 요청 코드(PK) 생성 
 		String latestOcdCd = requestMapper.findLatestOcdCd();
 		int nextNum = (latestOcdCd == null) ? 1 : Integer.parseInt(latestOcdCd.substring(5))+1;
 		String newOcdCd = String.format("OCD_C%05d",nextNum);
-		outsourcingRequest.setOcdCd(newOcdCd);
+		request.setOcdCd(newOcdCd);
 		
+		// 2. 이 요청을 위한 content_list 레코드 설정
 		String newClCd = "LIST_"+ newOcdCd;
 		requestMapper.insertContentListForRequest(newClCd, newOcdCd);
-		outsourcingRequest.setClCd(newClCd);
+		request.setClCd(newClCd);
 		
-		outsourcingRequest.setOcdReqType("신청");
-		outsourcingRequest.setStcCd("SD_BEFORE");
+		// 3. 초기 상태 설정 '신청', '승인 전'으로 설정합니다.
+		request.setOcdReqType("신청");
+		request.setStcCd("SD_BEFORE");
 		
-		requestMapper.insertOutsourcingRequest(outsourcingRequest);
+		// 4. DB에 최종 저장
+		requestMapper.insertOutsourcingRequest(request);
 	}
 }
-*/
+
