+++ outpolic/src/main/resources/templates/enter/outsourcing/outsourcingRequestListView.html
@@ -20,60 +20,41 @@
             <h1 class="section-title">새 외주 프로젝트 신청</h1>
             
             <div class="card">
-                <form id="outsourcingApplicationForm">
-                    <!-- 신청하는 외주 프로젝트 (읽기 전용) -->
-                    <div class="form-group">
-                        <label for="outsourcing_title">신청하는 외주 프로젝트</label>
-                        <input type="text" id="outsourcing_title" class="form-control bg-gray-100 cursor-not-allowed" value="[선택된 외주 프로젝트 제목]" readonly>
-                        <p class="form-instruction">현재 신청하려는 외주 프로젝트의 제목입니다. 변경할 수 없습니다.</p>
-                        <!-- Hidden field to pass the project ID -->
-                        <input type="hidden" id="outsourcing_id" value="[아웃소싱_ID_값]">
-                    </div>
-
-                    <!-- 제안 상세 설명 -->
-                    <div class="form-group">
-                        <label for="proposal_description">제안 상세 설명 <span class="required-star">*</span></label>
-                        <textarea id="proposal_description" name="proposal_description" class="form-control" rows="8" placeholder="외주 프로젝트에 대한 제안 내용을 상세하게 작성해 주세요. (예: 어떤 방식으로 진행할지, 예상되는 결과물, 공급자님의 강점 등)" required></textarea>
-                        <p class="form-instruction">공급자에게 프로젝트 수행 방안과 강점을 명확히 어필해 주세요.</p>
-                    </div>
-
-                    <!-- 희망 작업 기간 (시작일 및 종료일) -->
-                    <div class="flex-row">
-                        <div class="form-group">
-                            <label for="desired_work_start_date">희망 작업 시작일</label>
-                            <input type="date" id="desired_work_start_date" class="form-control">
-                        </div>
-                        <div class="form-group">
-                            <label for="desired_work_end_date">희망 작업 종료일</label>
-                            <input type="date" id="desired_work_end_date" class="form-control">
-                        </div>
-                    </div>
-
-                    <!-- 제안 금액 (희망 금액) -->
-                    <div class="form-group">
-                        <label for="proposed_amount">제안 금액 (세전 기준) <span class="required-star">*</span></label>
-                        <div class="input-with-unit">
-                            <input type="number" id="proposed_amount" name="proposed_amount" class="form-control" placeholder="제안할 금액을 입력하세요." min="0" required>
-                            <span>원</span>
-                        </div>
-                        <p class="form-instruction">프로젝트를 수행할 금액을 제안합니다. (필요시 협의 가능)</p>
-                    </div>
-
-                    <!-- 참고 자료 (파일 첨부) -->
-                    <div class="form-group">
-                        <label for="application_reference_files">참고 자료 첨부</label>
-                        <input type="file" id="application_reference_files" name="application_reference_files" accept="image/*, .pdf, .doc, .docx, .ppt, .pptx" multiple class="form-control p-2">
-                        <p class="form-instruction">프로젝트 제안 관련 포트폴리오, 기획안, 레퍼런스 등 자료를 첨부해주세요. (최대 5개)</p>
-                        <div class="image-preview-container mt-3" id="application-file-preview-container">
-                            <!-- File previews (images directly, others as icons) will appear here -->
-                        </div>
-                    </div>
-                    
-                    <!-- 외주 프로젝트 신청 버튼 -->
-                    <button type="submit" class="btn-submit">
-                        <i class="fas fa-check-circle mr-2"></i> 외주 프로젝트 신청하기
-                    </button>
-                </form>
+              		 <form id="outsourcingApplicationForm" action="/enter/requests/add" method="post">
+    
+				    <input type="hidden" name="entCd" th:value="${supplierCd}"> <div class="form-group">
+				        <label for="ocdTtl">프로젝트 제목 <span class="required-star">*</span></label>
+				        <input type="text" id="ocdTtl" name="ocdTtl" class="form-control" required>
+				    </div>
+				
+				    <div class="form-group">
+				        <label for="ocdExpln">제안 상세 설명 <span class="required-star">*</span></label>
+				        <textarea id="ocdExpln" name="ocdExpln" class="form-control" rows="8" required></textarea>
+				    </div>
+				
+				    <div class="flex-row">
+				        <div class="form-group">
+				            <label for="ocdDedline">희망 작업 기한</label>
+				            <input type="date" id="ocdDedline" name="ocdDedline" class="form-control">
+				        </div>
+				        <div class="form-group">
+				            <label for="ocdAmt">제안 금액 (세전 기준) <span class="required-star">*</span></label>
+				            <div class="input-with-unit">
+				                <input type="number" id="ocdAmt" name="ocdAmt" class="form-control" required>
+				                <span>원</span>
+				            </div>
+				        </div>
+				    </div>
+				    
+				    <div class="form-group">
+				        <label for="application_reference_files">참고 자료 첨부</label>
+				        <input type="file" id="application_reference_files" name="files" multiple>
+				    </div>
+				    
+				    <button type="submit" class="btn-submit">
+				        <i class="fas fa-check-circle mr-2"></i> 외주 프로젝트 신청하기
+				    </button>
+				</form>
             </div>
         </div>
         </th:block>
