+++ outpolic/src/main/resources/templates/enter/portfolio/addPortfolio.html
@@ -161,23 +161,25 @@
 
             $('#portfolioAddForm').on('submit', function(e) {
                 e.preventDefault();
+                const form = $(this);
+                const submitBtn = form.find('button[type="submit"]'); // ✅ 버튼 요소를 미리 찾아둠
+
+                // ✅ 버튼을 비활성화해서 중복 제출을 막습니다.
+                submitBtn.prop('disabled', true).text('저장 중...');
+
                 const formData = new FormData(this);
 
-                // --- ▼▼▼ 핵심 수정 내용 ▼▼▼ ---
-                // 1. 화면에 표시된 선택된 카테고리 뱃지에서 ID를 모두 가져와 배열로 만듭니다.
+                // --- 카테고리 코드를 하나의 문자열로 합치는 로직 (서버 설정 변경이 불가능할 때 사용) ---
+                formData.delete('categoryCodes'); 
                 const categoryIds = [];
                 $('#selected-categories-area .selected-item-badge').each(function() {
                     categoryIds.push($(this).data('id'));
                 });
-
-                // 2. ID 배열을 쉼표로 구분된 하나의 문자열로 합칩니다.
                 const categoryCodesString = categoryIds.join(',');
-
-                // 3. 합쳐진 문자열을 'categoryCodes'라는 이름의 단일 데이터로 FormData에 추가합니다.
                 if (categoryCodesString) {
                     formData.append('categoryCodes', categoryCodesString);
                 }
-                // --- ▲▲▲ 여기까지 수정 ▲▲▲ ---
+                // --- 여기까지 ---
 
                 $.ajax({
                     url: '/enter/portfolio/add-ajax',
@@ -191,11 +193,15 @@
                             window.location.href = res.redirectUrl;
                         } else {
                             alert('등록 실패: ' + (res.message || '알 수 없는 오류'));
+                            // 등록 실패 시 버튼을 다시 활성화
+                            submitBtn.prop('disabled', false).text('등록하기');
                         }
                     },
                     error: function(xhr) {
                         const errorMsg = xhr.responseJSON ? xhr.responseJSON.message : '서버 통신에 실패했습니다.';
                         alert('오류가 발생했습니다: ' + errorMsg);
+                        // 오류 발생 시에도 버튼을 다시 활성화
+                        submitBtn.prop('disabled', false).text('등록하기');
                     }
                 });
             });
